move configuration to separate module
